first draft of play-again feature

+++ src/main/java/ch/uzh/ifi/hase/soprafs23/service/LobbyService.java
@@ -177,6 +177,13 @@ public class LobbyService {
                     "Lobby needs at least 2 players to start the game");
         }
 
+        // check if lobby is collecting Play-Agains
+        if (lobby.isCollectingPlayAgains()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
+                    "Lobby is still collecting players for a re-match. Once the re-collecting time of 10 seconds" +
+                            " is over, the lobby creator can start the game again.");
+        }
+
         lobby.setJoinable(false);
         lobby = this.lobbyRepository.save(lobby);
         this.lobbyRepository.flush();
@@ -236,22 +243,30 @@ public class LobbyService {
                 this.gameService.sendLobbySettings(savedLobby.getLobbyId().intValue());
             }
             else {
-                log.info("Lobby {} has been deleted since last player left the lobby.",
-                        lobby.getLobbyId());
-                this.lobbyRepository.delete(lobby);
+                if (!lobby.isCollectingPlayAgains()) {
+                    log.info("Lobby {} has been deleted since last player left the lobby.",
+                            lobby.getLobbyId());
+                    this.lobbyRepository.delete(lobby);
+                }
             }
         }
 
-        player.setLobbyId(null);
-        player.setCreator(false);
-
-        this.playerRepository.save(player);
-        this.playerRepository.flush();
-
         this.lobbyRepository.flush();
 
+        this.playerService.clearLobbyConfigFromPlayer(playerToken);
 
         return savedLobby;
+    }
+
+    public synchronized void clearPlayerAfterGameEnd(String playerToken) {
+        Player player = this.playerService.getPlayerByToken(playerToken);
+        if (player == null) {
+            log.info("Player with playerToken {} is not a registered player." +
+                    "Returning without action taken.", playerToken);
+            return;
+        }
+
+        this.playerService.clearLobbyConfigFromPlayer(playerToken);
     }
 
     public synchronized void disconnectPlayer(String playerToken) {
@@ -265,7 +280,7 @@ public class LobbyService {
 
         Lobby lobby = getLobbyById(player.getLobbyId());
         if (lobby != null) {
-            log.info("Player {} is removed from lobby {} due to websocket disconnect", player.getPlayerName(), lobby.getLobbyId());
+            log.info("Player {} is removed from lobby {}.", player.getPlayerName(), lobby.getLobbyId());
 
             Lobby savedLobby = removePlayerFromLobby(player, lobby);
         }
@@ -274,6 +289,27 @@ public class LobbyService {
             log.info("Player {} is deleted since he is not a registered player.", player.getPlayerName());
 
             this.playerService.deletePlayer(player.getId(), player.getToken());
+        }
+
+    }
+
+    public void playAgain(Integer lobbyId, String wsConnectionId) {
+
+        Lobby lobby = getLobbyById(lobbyId);
+        Player requester = this.playerService.getPlayerByWsConnectionId(wsConnectionId);
+
+        if (requester != null && Objects.equals(lobbyId.longValue(), requester.getLobbyId())) {
+            if (lobby.isJoinable()) {
+
+                // first player that clicks "Play Again" becomes the new lobby creator
+                if (lobby.getJoinedPlayerNames().size() == 0) {
+                    lobby.setLobbyCreatorPlayerToken(requester.getToken());
+                }
+                lobby.addPlayerToLobby(requester.getPlayerName());
+                Lobby savedLobby = this.lobbyRepository.save(lobby);
+                this.lobbyRepository.flush();
+                this.gameService.sendLobbySettings(savedLobby.getLobbyId().intValue());
+            }
         }
 
     }
